reorganize infra folders
