+++ source/dotnet/wholesale-api/IntegrationEventPublishing/IntegrationEventPublishing.Application/UseCases/Factories/CompletedBatchFactory.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.Batches.Interfaces.Models;
-using Energinet.DataHub.Wholesale.IntegrationEventPublishing.Application.BatchAggregate;
+using Energinet.DataHub.Wholesale.IntegrationEventPublishing.Application.Model;
 using Energinet.DataHub.Wholesale.IntegrationEventPublishing.Application.UseCases.Mappers;
 using NodaTime.Extensions;
