+++ source/dotnet/wholesale-api/IntegrationEventPublishing/IntegrationEventPublishing.UnitTests/Application/UseCases/RegisterCompletedBatchesHandlerTests.cs
@@ -17,7 +17,7 @@ using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
 using Energinet.DataHub.Wholesale.Batches.Interfaces;
 using Energinet.DataHub.Wholesale.Batches.Interfaces.Models;
 using Energinet.DataHub.Wholesale.IntegrationEventPublishing.Application;
-using Energinet.DataHub.Wholesale.IntegrationEventPublishing.Application.BatchAggregate;
+using Energinet.DataHub.Wholesale.IntegrationEventPublishing.Application.Model;
 using Energinet.DataHub.Wholesale.IntegrationEventPublishing.Application.UseCases;
 using Energinet.DataHub.Wholesale.IntegrationEventPublishing.Application.UseCases.Mappers;
 using Moq;
