+++ source/dotnet/wholesale-api/IntegrationEventPublishing/IntegrationEventPublishing.UnitTests/Integration/QuantityQualityMapperTests.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
 using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model;
-using Energinet.DataHub.Wholesale.IntegrationEventPublishing.Infrastructure.Integration;
+using Energinet.DataHub.Wholesale.IntegrationEventPublishing.Infrastructure.IntegrationEvents.Mappers;
 using FluentAssertions;
 using Xunit;
