+++ source/dotnet/wholesale-api/WebApi.UnitTests/Infrastructure/InstantExtensionsTests.cs
@@ -13,6 +13,8 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.IntegrationEventPublishing.Infrastructure;
+using Energinet.DataHub.Wholesale.IntegrationEventPublishing.Infrastructure.IntegrationEvents;
+using Energinet.DataHub.Wholesale.IntegrationEventPublishing.Infrastructure.IntegrationEvents.Types;
 using FluentAssertions;
 using NodaTime;
 using Xunit;
